agents/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
pydantic>=2.10.0
azure-cosmos>=4.9.0
azure-functions>=1.21.0
azure-functions-durable>=1.2.0
azure-identity>=1.19.0
autogen-agentchat>=0.4.0
autogen-ext[openai]>=0.4.0
openai>=1.60.0
